requirements.txt
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
tavily-python>=0.3.0